==> RMD/8_human_epithelial_Krt6a.Rmd <==
---
title: "Human Epi Clusters Krt6a"
output: html_document
date: "2023-01-09"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Human and mouse epithelial subsets
Check for LY6D/KRT6A and Beta defensin expression in the mouse and human epithelial cells
```{r}

# Libraries
library(Seurat)
library(ggplot2)
library(viridis)
library(dplyr)

# set working directory
wrkdir <- "C:/Users/mgraha21/OneDrive - Johns Hopkins/Mindy JHMI Onedrive/MYC prostate scRNAseq/Manuscript Version of Analysis/Github uploads" #update specific to user
setwd(wrkdir)
```

# load datasets
```{r}
# human
human_seurat_epi <- readRDS("seurat_epi.rds")
DimPlot(human_seurat_epi, label = TRUE, raster = TRUE, shuffle = TRUE, group.by = "epithelial_groups", label.size = 5) + NoLegend() + ggtitle("Human Epithelial Clusters") + theme(plot.title = element_text(hjust = 0.5))

# mouse
mouse_seurat_epi <- readRDS(file = "mouse_seurat_epi.rds")
DimPlot(mouse_seurat_epi, label = TRUE,raster = TRUE, shuffle = TRUE, group.by = "Epi_clusters", label.size = 5) + NoLegend() + ggtitle("Mouse Epithelial Clusters") + theme(plot.title = element_text(hjust = 0.5))
```

# LY6D and KRT6A Expression
```{r}
# mouse
FeaturePlot(mouse_seurat_epi, features = c("Krt6a", "Ly6d"), order = T, min.cutoff = "q1",max.cutoff = "q99", blend = T, cols = c("light grey", "#219ebc", "#ffb703"), blend.threshold = 0, raster = FALSE) & NoAxes()

# human
FeaturePlot(human_seurat_epi, features = c("KRT6A", "LY6D"), order = T, min.cutoff = "q1",max.cutoff = "q99", blend = T, cols = c("light grey", "#219ebc", "#ffb703"), blend.threshold = 0, raster = FALSE) & NoAxes()
```

# beta defensins
```{r}
library(wesanderson)
pal <- wes_palette("Zissou1", n = 5, type = "continuous")

# human
FeaturePlot(human_seurat_epi, features = c("DEFB1"), order = T, raster = FALSE) + NoAxes() + scale_color_gradientn(colors = pal) + NoAxes()

# mouse
 FeaturePlot(mouse_seurat_epi, features = c("Defb29"), order = T, raster = FALSE) + NoAxes() + scale_color_gradientn(colors = pal) + NoAxes()
```
