==> RMD/14_human_immune.Rmd <==
---
title: "Human immune"
author: "Mindy Kim Graham"
date: "4/21/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Human Subjects Aggregated Prostatectomy Study
This analysis focuses on the immune susbset of the human dataset 
```{r}
# Libraries
library(Seurat)
library(ggplot2)
library(sctransform)
library(glmGamPoi)
library(RColorBrewer)
library(ggsci)
library(viridis)
library(dplyr)

# set working directory
wrkdir <- "C:/Users/mgraha21/OneDrive - Johns Hopkins/Mindy JHMI Onedrive/MYC prostate scRNAseq/Manuscript Version of Analysis/Github uploads" #update specific to user
setwd(wrkdir)
```

load data
```
# load seurat object
seurat_object <- readRDS(file = "seurat_prostatectomy.rds")

Idents(seurat_object) <- "cell_type"
DimPlot(seurat_object)
```

# subset to immune cells, sctransform, dimensionality reduction
refer to https://satijalab.org/seurat/articles/sctransform_vignette.html
```

# subset for immune cells
levels(seurat_object)
seurat_immune <- subset(seurat_object, idents = c("T cells", "Macrophages", "Mast cells", "B cells"))
DimPlot(seurat_immune)

# sctransform
seurat_immune <- SCTransform(seurat_immune, method = "glmGamPoi", verbose = FALSE, return.only.var.genes = FALSE, vst.flavor = "v2")

# dimensionality reduction
seurat_immune <- RunPCA(seurat_immune, verbose = FALSE)
seurat_immune <- RunUMAP(seurat_immune, reduction = "pca", dims = 1:40)
seurat_immune <- FindNeighbors(seurat_immune, dims = 1:40, verbose = FALSE)

DimPlot(seurat_immune, group.by = "cell_type")
DimPlot(seurat_immune, group.by = "Subject", shuffle = TRUE)

# cluster resolution
seurat_immune <- FindClusters(seurat_immune, verbose = FALSE, resolution = 0.2)
DimPlot(seurat_immune, label = TRUE, group.by = "SCT_snn_res.0.2", shuffle = TRUE)

```

cell type markers
```{r}

# immune subtypes
Immune <- c("PTPRC")

#myeloid
Mast <- c("TPSAB1", "TPSB2", "KIT")
Macrophages <- c("CD68", "C1QA", "C1QB", "C1QC")#HLA-DR+, CD11c/ITGAX-
myeloid <- c("ITGAM")
neutrophils <- c("FCGR3A", "FCGR3B", "CEACAM8", "FUT4")
eosinophils <- c(	"CCR3", "SIGLEC8") # CD16- (FCGR3A/B)
basophils <- c("FCER1A") #CD117-(KIT)
basophils_activated <- c("CD63", "ENPP3") #CD63, CD203c/ENPP3
mast_cells <- c("KIT", "TPSAB1", "TPSB2")
mono_MDSC <- c("CD14") #CD15-/FUT4, HLA-DR-
PMN_MDSC <- c("FUT4") #CD14-, HLA-DR-
plasmacytoid_dendritic <- c("IL3RA") #CD123, HLA-DR+
dendritic <- c("ITGAX") #HLA-DR+, ITGAX/CD11c
dendritic_1 <- c("XCR1", "CLEC9A")
dendritic_2 <- c("CD1C", "SIRPA")
M1 <- c("CD86", "CD80", "NOS2") #CD86, CD80, iNOS/NOS2
M2 <- c("CD163", "MRC1") #CD163, CD206/MRC1
```

```
# myeloid heatmap
DoHeatmap(subset(seurat_immune, downsample = 50), 
          features = unique(c(Macrophages, myeloid, neutrophils, eosinophils, 
                              basophils, mast_cells, mono_MDSC, PMN_MDSC,
                              plasmacytoid_dendritic, dendritic, dendritic_1, dendritic_2,
                               M1, M2))) + 
  scale_fill_viridis()
```

```{r}
#lymphoid
Tcells <- c("CD2", "CD3G", "CD3D", "CD3E")
Tcells_CD8 <- c("CD8A", "CD8B")
Tcells_CD4 <- c("CD4")
Tcells_Th1 <- c("TBX21", "IFNG") #T-Bet, IFNgamma
Tcells_Th2 <- c("GATA3", "IL4")
Tcells_Th17 <- c("RORC", "IL17A", "IL17B", "IL17C", "IL17D", "IL25", "IL17F")#RORyt, IL17
Tcells_reg <- c("FOXP3", "IL2RA", "IKZF2", "CCR4", "CTLA4") #IL2RA/CD25
Tcells_Th9 <- c("SPI1", "IL9")# PU.1/SPI1, IL9
Tcells_Tfh <- c("BCL6", "CXCR5", "IL21")
Tcells_Th22 <- c("AHR", "IL22")
Tcells_activated <- c("CD69", "IL2RA") #CD69, CD25
Tcells_cytotoxic <- c("GZMA", "GZMB", "GZMK", "GZMM", "GZMH", "PRF1") #granzyme, perforin
Tcells_naive_CM <- c("PTPRC", "SELL", "CCR7") #CD45RA, CD62L, CCR7, Naive and CM same genes
#Tcells_effectoryMem PTPRC/CD45RO+,  SELL-, CCR7-
Tcells_exhausted_progenitor <- c("HNF1A", "TCF7", "PDCD1", "HAVCR2") #TCF1/TCF7+, PD1 Hi, TIM-3 High
Tcells_exhuasted_terminally <- c("TOX", "TOX2", "PDCD1", "TIGIT") #Tox/Tox2, PD1 high, TIGIT

NKT <- c("NCAM1") #CD56
NKcells <- c("KLRD1", "NKG7", "GNLY")

Bcells <-c("JCHAIN", "CD79A")
Bcells_naive <- c("IGHD") #IgD+/CD27-
Bcells_switched_memory <- c("CD27") #IgD-/CD27+#unswitched IgD+/CD27+

Plasma_cells <- c("TNFRSF17", "SDC1")#BCMA, CD138
```

```
# lymphoid heatmap
DoHeatmap(subset(seurat_immune, downsample = 50), 
          features = unique(c(Tcells, Tcells_CD4, Tcells_CD8,Tcells_cytotoxic, Tcells_helper,
                              Tcells_Th1, Tcells_Th2, Tcells_Th17,
                              Tcells_reg, Tcells_Th9, Tcells_Tfh,
                              Tcells_Th22, Tcells_activated, Tcells_cytotoxic,
                              Tcells_naive_CM, Tcells_exhausted_progenitor, Tcells_exhuasted_terminally,
                              NKT, Bcells, Bcells_naive, Bcells_switched_memory, Plasma_cells, NKcells))) +
  scale_fill_viridis() 

# generate cell type heatmap
DoHeatmap(subset(seurat_object, downsample = 100), features = celltype_genes) +
  scale_fill_viridis()
```

# update cluster names with cell type
0 T cell
1 T cell
2 Macrophages
3 NKT cells
4 Mast cells
5 B cells
6 Macrophages
7 Plasma cells
8 pDCs

```
# epithelial groups
Idents(seurat_immune) <- "SCT_snn_res.0.2"
new.cluster.ids <- c("T cells",
                     "T cells", "Macrophages", "NKT cells", "Mast cells", "B cells",
                     "Macrophages", "Plasma cells", "pDCs")
names(new.cluster.ids) <- levels(seurat_immune)
seurat_immune[["immune"]] <- new.cluster.ids[as.numeric(as.matrix(seurat_immune@meta.data$SCT_snn_res.0.2))+1]
seurat_immune <- RenameIdents(seurat_immune, new.cluster.ids)

saveRDS(seurat_immune, file = "Humanseurat_immune.rds")
```

```{r}
# load immune seurat
seurat_immune <- readRDS(file = "Humanseurat_immune.rds")
Idents(seurat_immune) <- "immune"
levels(seurat_immune) <- c("T cells", "NKT cells", "B cells", "Plasma cells", "Macrophages", "Mast cells", "pDCs")

DimPlot(seurat_immune, reduction = "umap", label = TRUE, pt.size = 2, shuffle = TRUE, raster = TRUE) + 
  NoLegend() + ggtitle("Human Immune")+
  theme(plot.title = element_text(hjust = 0.5))

DoHeatmap(subset(seurat_immune, downsample = 50), 
          features = unique(c(Tcells, NKcells, Bcells, Plasma_cells, Macrophages, mast_cells, plasmacytoid_dendritic))) + 
  scale_fill_viridis()

```

#Treg
```{r}
library(UCell)
library(wesanderson)

genelist <- list()
genelist$Treg <- c("IL2RA", "FOXP3", "CD4", "IKZF2", "CCR4", "CTLA4")
seurat_immune<- AddModuleScore_UCell(seurat_immune, features = genelist)

pal <- wes_palette("Zissou1", type = "continuous", n =5)

FeaturePlot(seurat_immune, features = "Treg_UCell", order = TRUE, pt.size = 1) + scale_color_gradientn(colours = pal)
```

# Macrophage TREM2
```{r}
DimPlot(seurat_immune, group.by = "immune", label = TRUE, label.size = 6, raster = TRUE) + NoLegend() + ggtitle("Human Immune Clusters")
FeaturePlot(seurat_immune, features = "TREM2", order = TRUE, pt.size = 1)+ scale_color_gradientn(colours = pal)
```

# save seurat object
```
saveRDS(seurat_immune, file = "Humanseurat_immune.rds")
```
